requirements.txt
pytorch_lightning
torchmetrics
aiohttp
torch
numpy
pillow
torchvision
tqdm
humanfriendly
huggingface_hub
datasets
mmh3
timm
